+++ website/src/containers/SwapPanel/index.jsx
@@ -15,7 +15,7 @@ import { QuotePanel } from '../QuotePanel';
 import YouWillGet from '../../components/Bits/Youwillget';
 
 export const SwapPanel = () => {
-  const useIvan = true;
+  const useIvan = false;
   const { fromToken, txAmount, toToken } = useActions();
   const { quoteValid, setQuote, toTokenAmount } = useQuote();
